// apps/api-server/src/modules/auth/entities/Permission.d.ts
import type { Role } from './Role.js';
export declare class Permission {
    id: string;
    key: string;
    description: string;
    category: string;
    appId?: string;
    isActive: boolean;
    createdAt: Date;
    updatedAt: Date;
    roles?: Role[];
    static parseKey(key: string): {
        category: string;
        action: string;
    };
    getCategory(): string;
    getAction(): string;
}
//# sourceMappingURL=Permission.d.ts.map